packages/vinefeeder/src/vinefeeder/requirements.txt
beaupy
click
httpx
jmespath
PyQt6
PyQt6_sip
PyYAML
rich
Scrapy
